feat: Visualize shape combinations and save as PNG (master)
This commit introduces the ability to visualize combinations
of shapes and save the visualizations as PNG images.  It
also refactors the analysis script to save results in a
more structured JSON format, simplifying the data
representation for combinations of the second non-flag
column with other columns.  The visualization component
provides a visual representation of these combinations,
enhancing the analysis process.  Additionally, the
calculation and output of the compression ratio have been
removed, and combination analysis between the first two
non-flag columns is no longer performed.  The .gitignore
file is updated to exclude generated PNG images.

Signed-off-by: Aleksander Augustyn <aleksander.augustyn@ncbj.gov.pl>