``` feat: Visualize shape combinations and save as PNG (master)
This commit introduces the ability to visualize
combinations of shapes and save them as PNG images. It
refactors the analysis script to save results in a
structured JSON format, simplifying data
representation for combinations of the second non-flag
column with others. The visualization component
provides a visual representation of these
combinations, enhancing analysis.  Calculation and
output of the compression ratio have been removed,
and combination analysis between the first two
non-flag columns is no longer performed. The
.gitignore file is updated to exclude generated PNG
images.
```

Signed-off-by: Aleksander Augustyn <aleksander.augustyn@ncbj.gov.pl>