feat: Calculate volume of deformed nuclei (master)
This commit introduces the `BetaShapeCalculator.py`
module, which provides functions for calculating the
volume of deformed nuclei. It includes functions
for calculating the volume of a spherical nucleus,
the volume of a deformed nucleus using an
analytical formula, a volume fixing factor to
conserve volume, the nuclear radius as a function
of the polar angle, and a function to check if a
nucleus has a neck.  The `ParquetShapeSpaceAnalyzer`
no longer calculates or outputs a compression ratio
and removes the combination analysis between the
first two non-flag columns.

Signed-off-by: Aleksander Augustyn <aleksander.augustyn@ncbj.gov.pl>